+++ Backend/GameStoreApp.Application/Services/GenreService.cs
@@ -9,14 +9,14 @@ using System.Threading.Tasks;
 
 namespace GameStoreApp.Application.Services
 {
-    public class GenreService
+    public class GenreService : IGenreService
     {
         private readonly IGenreRepository _genreRepository;
         public GenreService(IGenreRepository genreRepository)
         {
             _genreRepository = genreRepository;
         }
-        public async Task<IEnumerable<GenreDto>> GetAllAsync()//gets all genres, maps to dtop
+        public async Task<IEnumerable<GenreDto>> GetAllAsync()
         {
             var genres = await _genreRepository.GetAllAsync();
             return genres.Select(g => new GenreDto
@@ -36,26 +36,33 @@ namespace GameStoreApp.Application.Services
                 Name = genre.Name
             };
         }
-        public async Task AddAsync(GenreDto dto)//create new GENRE
+        public async Task<GenreDto> CreateAsync(GenreDto dto)//create new GENRE
         {
             var genre = new Genre
             {
                 Name = dto.Name
             };
             await _genreRepository.AddAsync(genre);
+            dto.Id = genre.Id;
+            return dto;
         }
-        public async Task UpdateAsync(int id, GenreDto dto)//updates a genre
+        public async Task<GenreDto?> UpdateAsync(int id, GenreDto dto)//updates a genre
         {
             var genre = await _genreRepository.GetByIdAsync(id);
             if (genre == null)
-                throw new Exception("Genre not found...");
+                return null;
 
             genre.Name = dto.Name;
             await _genreRepository.UpdateAsync(genre);
+            return dto;
         }
-        public async Task DeleteAsync(int id)//deletes a genre
+        public async Task<bool> DeleteAsync(int id)//deletes a genre
         {
+            var genre = await _genreRepository.GetByIdAsync(id);
+            if (genre == null)
+                return false;
             await _genreRepository.DeleteAsync(id);
+            return true;
         }
     }
 }
